Refactor AgentDaemon
Ommision of AgentDaemon in favor of AgentDaemonMixin that creates
related run method that needs to be overrified from Daemon superclass.
BaseAgentDaemon and GracefulAgentDaemon created to facilitiy normal
agent daemonization and graceful daemon respectively.

AgentCLI updated to facilitate the choice between whether the daemon
needed can be a standard daemon or daemon that allows for graceful
stop/restart.