Daemon test_start completed
Created new python executable script to allow for daemon start method to
be tested properly. The daemon start method forks process and exits the
parent process, this causes for a SystemExit exception to occur if
called directly in test_start method.

Minor:

test_status updated to use test start script to emulate the running of
daemon.