Refactor daemon.py
Daemon stop method had a mechanism that would be trapped in an infinite
loop in the event that a lockfile existed. This mechanism was ommitted
to allow for the process to end no matter the state of a lock file. This
functinoality is addressed in the GracefulDaemon to make this feature
optional.